# Gex.NetCore/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Gex",
    "description": "Colleción de API's correspondientes al sistema de exámenes Gex.",
    "contact": {
      "name": "Agustin Ibañez",
      "email": "admin@gexsystem.com"
    },
    "version": "v1"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectGexResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectGexResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectGexResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectGexResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectGexResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectGexResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/comision": {
      "get": {
        "tags": [
          "Comision"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ComisionResponseGexResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComisionResponseGexResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComisionResponseGexResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Comision"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComisionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ComisionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ComisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ComisionResponseGexResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComisionResponseGexResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComisionResponseGexResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ComisionResponseGexResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComisionResponseGexResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComisionResponseGexResult"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Comision"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComisionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ComisionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ComisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ComisionResponseGexResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComisionResponseGexResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComisionResponseGexResult"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Comision"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ComisionResponseGexResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComisionResponseGexResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComisionResponseGexResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/comision/{id}": {
      "get": {
        "tags": [
          "Comision"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ComisionResponseGexResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComisionResponseGexResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComisionResponseGexResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/comision/nombre/{nombre}": {
      "get": {
        "tags": [
          "Comision"
        ],
        "parameters": [
          {
            "name": "nombre",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ComisionResponseGexResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComisionResponseGexResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComisionResponseGexResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "LoginRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "maxLength": 255,
            "minLength": 6,
            "type": "string",
            "format": "password"
          }
        },
        "nullable": true
      },
      "Object": {
        "type": "object",
        "nullable": true
      },
      "GexErrorMessage": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9,
          10
        ],
        "type": "integer",
        "format": "int32"
      },
      "ObjectGexResult": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Object"
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "error_messages": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string",
                "nullable": true
              },
              "nullable": true
            },
            "nullable": true
          },
          "error_code": {
            "$ref": "#/components/schemas/GexErrorMessage"
          }
        },
        "nullable": true
      },
      "RegistroRequest": {
        "required": [
          "confirm_password",
          "email",
          "first_name",
          "last_name",
          "location",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "maxLength": 255,
            "minLength": 6,
            "type": "string",
            "format": "password"
          },
          "confirm_password": {
            "maxLength": 255,
            "minLength": 6,
            "type": "string",
            "format": "password"
          },
          "first_name": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "last_name": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "location": {
            "maxLength": 150,
            "minLength": 0,
            "type": "string"
          }
        },
        "nullable": true
      },
      "ComisionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "ciclo_lectivo": {
            "type": "integer",
            "format": "int32"
          }
        },
        "nullable": true
      },
      "ComisionResponseGexResult": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ComisionResponse"
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "error_messages": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string",
                "nullable": true
              },
              "nullable": true
            },
            "nullable": true
          },
          "error_code": {
            "$ref": "#/components/schemas/GexErrorMessage"
          }
        },
        "nullable": true
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "extensions": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/Object"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": {
          "$ref": "#/components/schemas/Object"
        },
        "nullable": true
      },
      "ComisionRequest": {
        "required": [
          "ciclo_lectivo",
          "id",
          "nombre"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "maxLength": 50,
            "minLength": 5,
            "type": "string"
          },
          "ciclo_lectivo": {
            "maximum": 2022,
            "minimum": 2001,
            "type": "integer",
            "format": "int32"
          }
        },
        "nullable": true
      },
...
    }
  }
}